+++ sdk/compose/icons/src/commonMain/kotlin/io/github/composegears/valkyrie/sdk/compose/icons/colored/TablerLogo.kt
@@ -1,0 +1,84 @@
+package io.github.composegears.valkyrie.sdk.compose.icons.colored
+
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.SolidColor
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.graphics.vector.path
+import androidx.compose.ui.unit.dp
+import io.github.composegears.valkyrie.sdk.compose.icons.ValkyrieIcons
+
+val ValkyrieIcons.Colored.TablerLogo: ImageVector
+    get() {
+        if (_TablerLogo != null) {
+            return _TablerLogo!!
+        }
+        _TablerLogo = ImageVector.Builder(
+            name = "TablerLogo",
+            defaultWidth = 48.dp,
+            defaultHeight = 48.dp,
+            viewportWidth = 48f,
+            viewportHeight = 48f,
+        ).apply {
+            path(fill = SolidColor(Color(0xFF216CC4))) {
+                moveTo(23.813f, -0.25f)
+                curveTo(24.989f, -0.271f, 26.165f, -0.291f, 27.377f, -0.313f)
+                curveTo(38.246f, -0.362f, 38.246f, -0.362f, 43f, 3f)
+                curveTo(48.938f, 9.419f, 48.214f, 15.488f, 48.25f, 23.813f)
+                curveTo(48.271f, 24.989f, 48.291f, 26.165f, 48.313f, 27.377f)
+                curveTo(48.362f, 38.246f, 48.362f, 38.246f, 45f, 43f)
+                curveTo(38.581f, 48.938f, 32.512f, 48.214f, 24.188f, 48.25f)
+                curveTo(23.011f, 48.271f, 21.835f, 48.291f, 20.623f, 48.313f)
+                curveTo(9.754f, 48.362f, 9.754f, 48.362f, 5f, 45f)
+                curveTo(-0.938f, 38.581f, -0.214f, 32.512f, -0.25f, 24.188f)
+                curveTo(-0.271f, 23.011f, -0.291f, 21.835f, -0.313f, 20.623f)
+                curveTo(-0.362f, 9.754f, -0.362f, 9.754f, 3f, 5f)
+                curveTo(9.419f, -0.938f, 15.488f, -0.214f, 23.813f, -0.25f)
+                close()
+            }
+            path(fill = SolidColor(Color(0xFFF6F9FC))) {
+                moveTo(23.063f, 15.563f)
+                curveTo(26.731f, 16.109f, 28.436f, 17.173f, 30.941f, 19.867f)
+                curveTo(33.133f, 24.283f, 31.834f, 29.312f, 31f, 34f)
+                curveTo(29.751f, 34.482f, 28.501f, 34.96f, 27.25f, 35.438f)
+                curveTo(26.554f, 35.704f, 25.858f, 35.971f, 25.141f, 36.246f)
+                curveTo(20.624f, 37.837f, 16.815f, 38.093f, 12f, 38f)
+                curveTo(15.55f, 33.405f, 18.477f, 30.235f, 24f, 28f)
+                curveTo(23.67f, 26.68f, 23.34f, 25.36f, 23f, 24f)
+                curveTo(21.041f, 24.113f, 21.041f, 24.113f, 19f, 25f)
+                curveTo(18.732f, 25.722f, 18.464f, 26.444f, 18.188f, 27.188f)
+                curveTo(16.474f, 31.246f, 13.42f, 33.339f, 10f, 36f)
+                curveTo(10.516f, 29.501f, 11.467f, 23.338f, 13f, 17f)
+                curveTo(16.348f, 16.464f, 19.707f, 16.042f, 23.063f, 15.563f)
+                close()
+            }
+            path(fill = SolidColor(Color(0xFF377ACA))) {
+                moveTo(20f, 23f)
+                curveTo(22.063f, 23.438f, 22.063f, 23.438f, 24f, 24f)
+                curveTo(24.625f, 25.875f, 24.625f, 25.875f, 25f, 28f)
+                curveTo(23f, 30f, 23f, 30f, 20.75f, 30.813f)
+                curveTo(17.202f, 32.345f, 14.892f, 34.444f, 12f, 37f)
+                curveTo(12f, 34f, 12f, 34f, 14.375f, 31.625f)
+                curveTo(15.241f, 30.759f, 16.108f, 29.892f, 17f, 29f)
+                curveTo(18.254f, 26.368f, 18.254f, 26.368f, 19f, 24f)
+                curveTo(19.33f, 23.67f, 19.66f, 23.34f, 20f, 23f)
+                close()
+            }
+            path(fill = SolidColor(Color(0xFFF1F5FB))) {
+                moveTo(27f, 10f)
+                curveTo(30.158f, 11.398f, 32.323f, 13.052f, 34.75f, 15.5f)
+                curveTo(35.67f, 16.413f, 35.67f, 16.413f, 36.609f, 17.344f)
+                curveTo(38f, 19f, 38f, 19f, 38f, 21f)
+                curveTo(36.27f, 21.73f, 36.27f, 21.73f, 34f, 22f)
+                curveTo(31.824f, 20.488f, 31.824f, 20.488f, 29.688f, 18.313f)
+                curveTo(28.969f, 17.607f, 28.251f, 16.902f, 27.512f, 16.176f)
+                curveTo(26f, 14f, 26f, 14f, 26.27f, 11.73f)
+                curveTo(26.511f, 11.159f, 26.752f, 10.588f, 27f, 10f)
+                close()
+            }
+        }.build()
+
+        return _TablerLogo!!
+    }
+
+@Suppress("ObjectPropertyName")
+private var _TablerLogo: ImageVector? = null